// src/enums/slot-symbols.enum.ts
export enum slotSymbols {
    s9 = '9',
    s10 = '10',
    sA = 'A',
    sH1 = 'H1',
    sH2 = 'H2',
    sH3 = 'H3',
    sH4 = 'H4',
    sH5 = 'H5',
    sH6 = 'H6',
    sJ = 'J',
    sK = 'K',
    sM1 = 'M1',
    sM2 = 'M2',
    sM3 = 'M3',
    sM4 = 'M4',
    sM5 = 'M5',
    sM6 = 'M6',
    sQ = 'Q',
}